// csm.web.app/src/app/shared/alert/enums/alert-type.enum.ts
/*
 * *************************************************************************
 *
 * Copyright: Robert Bosch Power Tools GmbH, 2020
 *
 * *************************************************************************
 */

export enum AlertTypeEnum {
    Error = 'error',
    Success = 'success',
    Warning = 'warning',
    Neutral = 'neutral',
}